Add LineController to divide responsibilities from StationController for single responsibility principle
